requirements.txt
anyio==4.10.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
exceptiongroup==1.3.0
firebase_admin==7.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
msgpack==1.1.1
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
